Moved all IP-related classes to its own package.
