+++ src/main/java/com/craftfire/commons/ip/IPRange.java
@@ -1,4 +1,4 @@
-package com.craftfire.commons.classes;
+package com.craftfire.commons.ip;
 
 public interface IPRange {
     boolean isInRange(IPAddress address);
