+++ src/main/java/com/craftfire/commons/ip/IPv6Address.java
@@ -1,4 +1,4 @@
-package com.craftfire.commons.classes;
+package com.craftfire.commons.ip;
 
 import java.net.Inet6Address;
 import java.net.InetAddress;
